feat(js): added ios setup partially
