+++ wrappers/javascript/react-native/cpp/indyVdr.cpp
@@ -2,9 +2,11 @@
 
 namespace indyVdr {
 
-const char* getVersion() {
-  const char* version = ::indy_vdr_version();
-  return version;
+jsi::Value getVersion(jsi::Runtime &rt, jsi::Object options) {
+  const char *version = indy_vdr_get_version();
+  jsi::Object object = jsi::Object(rt);
+  object.setProperty(rt, "version", version);
+  return object;
 }
 
-}
+} // namespace indyVdr
