+++ wrappers/javascript/react-native/ios/IndyVdrReactNative.h
@@ -1,11 +1,5 @@
 #import <React/RCTBridgeModule.h>
 
-#ifdef __cplusplus
-
-#import "react-native-indy-vdr-react-native.h"
-
-#endif
-
 @interface IndyVdrReactNative : NSObject <RCTBridgeModule>
 
 @end
